requirements.txt
aiogram==3.8.0
psycopg2-binary==2.9.9
python-dotenv==1.0.1
asyncpg==0.29.0
pandas
xlsxwriter
